Move email display toggle from main toolbar to email action buttons
