+++ app/internal_packages/thread-list/lib/main.ts
@@ -8,13 +8,7 @@ import MessageListToolbar from './message-list-toolbar';
 import SelectedItemsStack from './selected-items-stack';
 import * as ThreadPermalinkHandler from './thread-permalink-handler';
 
-import {
-  UpButton,
-  DownButton,
-  MoveButtons,
-  FlagButtons,
-  ToggleEmailRenderModeButton,
-} from './thread-toolbar-buttons';
+import { UpButton, DownButton, MoveButtons, FlagButtons } from './thread-toolbar-buttons';
 
 export function activate() {
   ThreadPermalinkHandler.activate();
@@ -66,10 +60,6 @@ export function activate() {
   ComponentRegistry.register(FlagButtons, {
     role: 'ThreadActionsToolbarButton',
   });
-
-  ComponentRegistry.register(ToggleEmailRenderModeButton, {
-    role: 'MessageList:ThreadActionsToolbarButton',
-  });
 }
 
 export function deactivate() {
@@ -79,7 +69,6 @@ export function deactivate() {
   ComponentRegistry.unregister(MessageListToolbar);
   ComponentRegistry.unregister(MoveButtons);
   ComponentRegistry.unregister(FlagButtons);
-  ComponentRegistry.unregister(ToggleEmailRenderModeButton);
   ComponentRegistry.unregister(UpButton);
   ComponentRegistry.unregister(DownButton);
 }
